Adds completion for first command of plugins
This commit allows the user to get bash/zsh completion of the first
command of a plugin.

For example, with the use of a plugin named 'kubectl-env', completing
   kubectl e<TAB>
will include 'env' in the list of possible completions.

Because bash completion is generated by Cobra, this commit took the
approach of registering every plugin as a Cobra command.  This makes
Cobra include each plugin command into the completion script.
For efficiency, searching for plugins and registering them as Cobra
commands is only done when running the command 'kubectl completion',
which is the only time we need it.